Rename server apis
